+++ src/domain/dtos/FileUploadOptions.ts
@@ -58,4 +58,10 @@ export class FileUploadOptions {
         nullable: true,
     })
     fileType: string;
+
+    @IntegerField({
+        label: 'User ID',
+        nullable: true,
+    })
+    userId?: number;
 }