+++ src/domain/services/FileImageService.ts
@@ -1,6 +1,8 @@
 import * as sharp from 'sharp';
 import {DataMapper} from '@steroidsjs/nest/usecases/helpers/DataMapper';
 import {ContextDto} from '@steroidsjs/nest/usecases/dtos/ContextDto';
+import {Inject, Injectable} from '@nestjs/common';
+import {EventEmitter2} from '@nestjs/event-emitter';
 import {IFileImageRepository} from '../interfaces/IFileImageRepository';
 import {FileImageModel} from '../models/FileImageModel';
 import {FileModel} from '../models/FileModel';
@@ -16,11 +18,16 @@ import {FileConfigService} from './FileConfigService';
 
 const SVG_MIME_TYPE = 'image/svg+xml';
 
+@Injectable()
 export class FileImageService {
     constructor(
+        @Inject(IFileImageRepository)
         public repository: IFileImageRepository,
+        @Inject(FileConfigService)
         protected readonly fileConfigService: FileConfigService,
+        @Inject(IFileStorageFactory)
         protected readonly fileStorageFactory: IFileStorageFactory,
+        @Inject(EventEmitter2)
         protected readonly eventEmitter: IEventEmitter,
     ) {
     }