+++ src/domain/validators/index.ts
@@ -1,0 +1,7 @@
+import {FileMaxSizeValidator} from './FileMaxSizeValidator';
+import {FileMimeTypesValidator} from './FileMimeTypesValidator';
+
+export const fileValidators = [
+    FileMaxSizeValidator,
+    FileMimeTypesValidator,
+];