+++ src/infrastructure/module.ts
@@ -1,14 +1,11 @@
-import {ModuleHelper} from '@steroidsjs/nest/infrastructure/helpers/ModuleHelper';
 import {IFileService} from '@steroidsjs/nest-modules/file/services/IFileService';
-import {EventEmitter2} from '@nestjs/event-emitter';
+import {IValidator} from '@steroidsjs/nest/usecases/interfaces/IValidator';
 import {ModuleRef} from '@nestjs/core';
 import {IFileRepository} from '../domain/interfaces/IFileRepository';
 import {IFileImageRepository} from '../domain/interfaces/IFileImageRepository';
 import {FileService} from '../domain/services/FileService';
 import {FileImageService} from '../domain/services/FileImageService';
 import {FileConfigService} from '../domain/services/FileConfigService';
-import {FileMaxSizeValidator} from '../domain/validators/FileMaxSizeValidator';
-import {FileMimeTypesValidator} from '../domain/validators/FileMimeTypesValidator';
 import {FileStorageFactory} from '../domain/services/FileStorageFactory';
 import {FileLocalStorage} from '../domain/storages/FileLocalStorage';
 import {MinioS3Storage} from '../domain/storages/MinioS3Storage';
@@ -18,6 +15,8 @@ import {FileRemovedEventHandleUseCase} from '../usecases/fileRemovedEventHandleU
 import {IFileTypeService} from '../domain/interfaces/IFileTypeService';
 import {FileTypeService} from '../domain/services/FileTypeService';
 import {IFileStorageFactory} from '../domain/interfaces/IFileStorageFactory';
+import {fileValidators} from '../domain/validators';
+import {FILE_VALIDATORS_TOKEN} from '../domain/constants/FileValidatorsToken';
 import {FILE_STORAGES_TOKEN} from '../domain/interfaces/IFileStorage';
 import {FileEventsSubscriber} from './subscribers/FileEventsSubscriber';
 import {CronJobsRegister} from './services/CronJobsRegister';
@@ -43,8 +42,12 @@ export default (config: IFileModuleConfig) => ({
         CronJobsRegister,
 
         // Validators
-        FileMaxSizeValidator,
-        FileMimeTypesValidator,
+        ...fileValidators,
+        {
+            provide: FILE_VALIDATORS_TOKEN,
+            useFactory: (...providers: IValidator[]) => providers,
+            inject: fileValidators,
+        },
 
         // Storages
         FileLocalStorage,
@@ -73,24 +76,14 @@ export default (config: IFileModuleConfig) => ({
             provide: IFileStorageFactory,
             useClass: FileStorageFactory,
         },
-        ModuleHelper.provide(FileService, IFileService, [
-            IFileRepository,
-            FileImageService,
-            FileConfigService,
-            IFileStorageFactory,
-            EventEmitter2,
-            IFileTypeService,
-            [
-                FileMimeTypesValidator,
-                FileMaxSizeValidator,
-            ],
-        ]),
-        ModuleHelper.provide(FileImageService, [
-            IFileImageRepository,
-            FileConfigService,
-            IFileStorageFactory,
-            EventEmitter2,
-        ]),
+        {
+            provide: IFileService,
+            useClass: FileService,
+        },
+        {
+            provide: FileImageService,
+            useClass: FileImageService,
+        },
 
         DeleteLostAndTemporaryFilesService,
 
@@ -98,10 +91,7 @@ export default (config: IFileModuleConfig) => ({
         FileEventsSubscriber,
 
         // UseCases
-        ModuleHelper.provide(FileRemovedEventHandleUseCase, [
-            IFileStorageFactory,
-            FileConfigService,
-        ]),
+        FileRemovedEventHandleUseCase,
         ClearUnusedFilesCommand,
     ],
     exports: [