+++ src/usecases/getFilePathModels/GetFileModelsPathUsecase.ts
@@ -1,4 +1,5 @@
 import {Inject, Injectable} from '@nestjs/common';
+import {IFileService} from '@steroidsjs/nest-modules/file/services/IFileService';
 import {FileService} from '../../domain/services/FileService';
 import {FileImageService} from '../../domain/services/FileImageService';
 import {IGetFileModelsPathUsecase} from './interfaces/IGetFileModelsPathUsecase';
@@ -6,9 +7,8 @@ import {IGetFileModelsPathUsecase} from './interfaces/IGetFileModelsPathUsecase'
 @Injectable()
 export class GetFileModelsPathUsecase implements IGetFileModelsPathUsecase {
     constructor(
-        @Inject(FileService)
+        @Inject(IFileService)
         protected readonly fileService: FileService,
-        @Inject(FileImageService)
         protected readonly fileImageService: FileImageService,
     ) {}